+++ examples/data-sources/onepasswordorg_vault/data-source.tf
@@ -1,0 +1,7 @@
+data "onepasswordorg_vault" "test" {
+  name = "test-vault"
+}
+
+output "vault_test" {
+  value = data.onepasswordorg_vault.test
+}